+++ src/main/java/com/moment/the/config/security/SecurityConfiguration.java
@@ -1,6 +1,6 @@
 package com.moment.the.config.security;
 
-import com.moment.the.config.mvc.ExceptionHandlerFilter;
+import com.moment.the.exception.handler.ExceptionHandlerFilter;
 import com.moment.the.config.security.jwt.JwtRequestFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
